+++ src/test/scala/ai/starlake/schema/handlers/SchemaHandlerSpec.scala
@@ -634,6 +634,30 @@ class SchemaHandlerSpec extends TestHelper {
       }
     }
 
+    "Load Business with jinja" should "should not run jinja parser" in {
+      new SpecTrait(
+        domainOrJobFilename = "locations.comet.yml",
+        sourceDomainOrJobPathname = "/sample/simple-json-locations/locations.comet.yml",
+        datasetDomainName = "locations",
+        sourceDatasetPathName = "/sample/simple-json-locations/locations.json"
+      ) {
+        import org.scalatest.TryValues._
+        cleanMetadata
+        cleanDatasets
+        val schemaHandler = new SchemaHandler(storageHandler)
+        val filename = "/sample/metadata/business/my-jinja-job.comet.yml"
+        val jobPath = new Path(getClass.getResource(filename).toURI)
+        val job = schemaHandler.loadJobFromFile(jobPath)
+        println(job)
+        job.success.value.tasks.head.sql.get.trim shouldBe """{% set myList = ["col1,", "col2"] %}
+                                                             |select
+                                                             |{%- for x in myList %}
+                                                             |{{x}}
+                                                             |{%- endfor %}
+                                                             |from dream_working.client""".stripMargin // Job renamed to filename and error is logged
+      }
+    }
+
     // TODO TOFIX
     //  "Load Business Definition" should "produce business dataset" in {
     //    val sh = new HdfsStorageHandler